+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog for NeoFS Node
 - Inability to restore RPC connection after the second disconnect (#2325)
 - Tree service panics when cleaning up failed connections (#2335)
 - Dropping small objects on any error on write-cache side (#2336)
+- Iterating over just removed files by FSTree (#98)
 
 ### Removed
 - Non-notary mode support for sidechain (#2321)